refactor(configurator): split UI page modules
